+++ src/app/src/screens/Machine/Navigation.ts
@@ -1,5 +1,5 @@
 import { pathTo } from "../../Navigator";
 
 export const getMachineUrl = (id: string, view: string) => {
-  return pathTo(`/screens/machines/${id}/${view}`);
+  return pathTo(`/screens/machines/${encodeURIComponent(id)}/${encodeURIComponent(view)}`);
 };