+++ src/app/src/screens/Secret/Navigation.ts
@@ -1,5 +1,5 @@
 import { pathTo } from "../../Navigator";
 
 export const getSecretUrl = (id: string, view: string) => {
-  return pathTo(`/screens/secrets/${id}/${view}`);
+  return pathTo(`/screens/secrets/${encodeURIComponent(id)}/${encodeURIComponent(view)}`);
 };