+++ src/app/src/screens/Settings/Navigation.ts
@@ -1,5 +1,5 @@
 import { pathTo } from "../../Navigator";
 
 export const getSettingsUrl = (view: string) => {
-  return pathTo(`/screens/settings/${view}`);
+  return pathTo(`/screens/settings/${encodeURIComponent(view)}`);
 };