+++ src/app/src/screens/Settings/UserSettingsScreen.tsx
@@ -1,16 +1,15 @@
-import { useCallback, useState } from "react";
-import { AnchorButton, Button, Callout, Checkbox, ControlGroup, FormGroup, HTMLSelect, Icon, InputGroup, Intent } from "@blueprintjs/core";
+import { useCallback } from "react";
+import { Button, Callout, Checkbox, ControlGroup, FormGroup, HTMLSelect } from "@blueprintjs/core";
 import { IconNames } from "@blueprintjs/icons";
 import { useTranslation } from "react-i18next";
 import * as ReactIcon from "@mdi/react";
 import { mdiEmoticonSad, mdiEmoticonWink } from "@mdi/js";
 
 // project
-import Environment, { LOGGING_LEVELS } from "../../Environment";
+import { LOGGING_LEVELS } from "../../Environment";
 import { AppScreen, AppScreenProps, UserConfigurationOptions } from "../../Types";
 import { ScreenHeader } from "./ScreenHeader";
 import { Native } from "../../Native";
-import { Notification } from "../../Notification";
 import { useStoreActions, useStoreState } from "../../domain/types";
 
 // module
@@ -27,10 +26,8 @@ export const View = "user-settings";
 export const Title = "Settings";
 
 export const Screen: AppScreen<ScreenProps> = () => {
-  const [programPaths, setProgramPaths] = useState<{ [key: string]: any }>({});
   const { t } = useTranslation();
   const pending = useStoreState((state) => state.pending);
-  const native = useStoreState((state) => state.native);
   const provisioned = useStoreState((state) => state.environment.provisioned);
   const system = useStoreState((state) => state.environment.system);
   const running = useStoreState((state) => state.environment.running);
@@ -38,44 +35,6 @@ export const Screen: AppScreen<ScreenProps> = () => {
   const connect = useStoreActions((actions) => actions.connect);
   const setUserConfiguration = useStoreActions((actions) => actions.setUserConfiguration);
   const program = userConfiguration.program;
-  const isValid = provisioned && program.currentVersion;
-  const onProgramSelectClick = useCallback(
-    async (e: React.MouseEvent<HTMLElement, MouseEvent>) => {
-      const sender = e.currentTarget as HTMLElement;
-      const field = sender?.closest(".AppSettingsField");
-      const program = field?.getAttribute("data-program-name");
-      const result = await Native.getInstance().openFileSelector();
-      if (result) {
-        const filePath = result?.filePaths[0];
-        if (!result.canceled && filePath && program) {
-          try {
-            const programSettings: any = {};
-            const programKey = `program.${userConfiguration.program.name}.path`;
-            programSettings[programKey] = filePath;
-            await setUserConfiguration(programSettings);
-            setProgramPaths((prev) => ({ ...prev, [program]: filePath }));
-          } catch (error) {
-            console.error("Unable to change program path", error);
-            Notification.show({ message: t("Unable to change program path"), intent: Intent.DANGER });
-          }
-        }
-      } else {
-        console.error("Unable to open file dialog");
-      }
-    },
-    [userConfiguration, setUserConfiguration, setProgramPaths, t]
-  );
-  const onProgramPathChange = useCallback(
-    (event: React.FormEvent<HTMLInputElement>) => {
-      const sender = event.currentTarget;
-      const field = sender?.closest(".AppSettingsField");
-      const program = field?.getAttribute("data-program-name");
-      if (program) {
-        setProgramPaths({ ...programPaths, [program]: sender.value });
-      }
-    },
-    [programPaths]
-  );
   const onConnectClick = useCallback(
     async () => {
       await connect({ startApi: true });
@@ -152,69 +111,6 @@ export const Screen: AppScreen<ScreenProps> = () => {
       <ScreenHeader currentScreen={ID} />
       <div className="AppScreenContent">
         {contentWidget}
-        <div className="AppSettingsForm" data-form="paths">
-          <div
-            className="AppSettingsField"
-            data-field="program.path"
-            data-program-name={program?.name}
-            data-program-present={isValid ? "yes" : "no"}
-          >
-            <div className="AppSettingUserConfigurationPath">
-              <Icon icon={IconNames.INFO_SIGN} />
-              <strong>{t('Application settings and logs path')}</strong>
-              <input type="text" value={userConfiguration.path} readOnly/>
-            </div>
-            <FormGroup
-              helperText={
-                <div className="AppSettingsFieldProgramHelper">
-                  &nbsp;
-                  {isValid ? (
-                    <span>{t("Detected version {{currentVersion}}", program)}</span>
-                  ) : program?.currentVersion ? (
-                    t("The location of the {{program}} executable binary", { program: program?.name })
-                  ) : (
-                    t("Could not detect current version")
-                  )}
-                </div>
-              }
-              label={
-                <AnchorButton
-                  minimal
-                  icon={isValid ? IconNames.THUMBS_UP : IconNames.THUMBS_DOWN}
-                  intent={isValid ? Intent.SUCCESS : Intent.DANGER}
-                  text={program.name}
-                  title={t("Go to {{name}} homepage", program)}
-                  target="_blank"
-                  href={program.homepage || ""}
-                />
-              }
-              labelFor={`${program.name}_path`}
-              labelInfo={t("(required)")}
-            >
-              <ControlGroup fill={true} vertical={false}>
-                <InputGroup
-                  fill
-                  id={`${program.name}_path`}
-                  readOnly={native}
-                  placeholder={"..."}
-                  value={programPaths[program.name || ""] || program.path}
-                  onChange={onProgramPathChange}
-                />
-                {native ? (
-                  <Button
-                    icon={IconNames.LOCATE}
-                    text={t("Select")}
-                    title={t("Select program")}
-                    intent={Intent.PRIMARY}
-                    onClick={onProgramSelectClick}
-                  />
-                ) : (
-                  <Button icon={IconNames.TICK} title={t("Accept")} />
-                )}
-              </ControlGroup>
-            </FormGroup>
-          </div>
-        </div>
         <ContainerEngineManager helperText={runningDetails} />
         <div className="AppSettingsForm" data-form="flags">
           <FormGroup