+++ src/app/src/screens/Volume/Navigation.ts
@@ -1,5 +1,5 @@
 import { pathTo } from "../../Navigator";
 
 export const getVolumeUrl = (id: string, view: string) => {
-  return pathTo(`/screens/volumes/${id}/${view}`);
+  return pathTo(`/screens/volumes/${encodeURIComponent(id)}/${encodeURIComponent(view)}`);
 };